# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /phone/lookup:
    get:
      operationId: phoneLookup
      tags: [phone]
      summary: Look up phone number intelligence
      description: >
        Runs an offline phone number intelligence lookup (line type, carrier,
        geolocation, fraud score, spam/DNC/reassigned checks). Requires an
        API key in the X-API-Key header.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: number
          in: query
          required: true
          schema:
            type: string
          example: "+14155552671"
          description: Phone number in international format
      responses:
        "200":
          description: Lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneLookupResult"
        "400":
          description: Missing or invalid number parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    AdminSecretAuth:
      type: apiKey
      in: header
      name: X-Admin-Secret
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    PhoneLookupResult:
      type: object
      description: >
        Phone intelligence result. Fields sourced from the phonenumbers library
        (valid, line_type, voip, carrier, country, city, region, timezones) are
        authoritative. Fields sourced from community datasets (spam, recent_abuse,
        spammer, dnc) reflect what has been reported. Fields requiring live
        carrier/breach/HLR access (name, associated_emails, user_activity,
        leaked_online, reassigned) are null — never fabricated.
      properties:
        # --- Authoritative (phonenumbers library) ---
        valid:
          type: boolean
          description: Whether the number is a valid, dialable number per ITU standards
        line_type:
          type: string
          description: "Fixed Line, Mobile, VoIP, Toll-Free, Premium Rate, etc."
        voip:
          type: boolean
          description: Whether the number is a VoIP line
        carrier:
          type: string
          description: >
            Original network carrier name (authoritative for most international numbers;
            may be empty for US mobile numbers due to number portability)
        country:
          type: string
          description: ISO 3166-1 alpha-2 country code
        city:
          type: string
          description: City or geographic area (from phonenumbers geocoder)
        region:
          type: string
          description: State, province, or region (from phonenumbers geocoder)
        timezones:
          type: array
          items:
            type: string
          description: IANA timezone identifiers for this number's geographic area
        # --- Community / heuristic ---
        active:
          type: boolean
          description: >
            Best-effort heuristic — true if number is valid and not flagged in any
            community abuse dataset. NOT a live reachability check.
        fraud_score:
          type: number
          description: Risk score 0–100 computed from community data and pattern analysis
        fraud_reasons:
          type: array
          items:
            type: string
          description: Reasons contributing to the fraud score
        recent_abuse:
          type: boolean
          description: True if number appears in community abuse/spam datasets
        spammer:
          type: boolean
          description: True if number appears in community spam datasets
        spam:
          type: boolean
          description: True if number appears in community spam datasets
        prepaid:
          type: boolean
          description: Heuristic — true if carrier name matches known prepaid brands
        risky:
          type: boolean
          description: True if fraud_score >= 75 or flagged in community datasets
        dnc:
          type: boolean
          description: >
            Community-proxy DNC flag — true if number appears in abuse reports
            (not the official FTC Do Not Call Registry, which requires subscription)
        dnc_source:
          type: string
          description: Source of DNC determination
        pattern_flags:
          type: array
          items:
            type: string
          description: Structural pattern risk indicators (e.g. 555-0xxx, sequential digits, high-risk area code)
        # --- Unavailable offline (always null) ---
        name:
          type: string
          nullable: true
          description: "Subscriber name — null: requires CNAM carrier lookup"
        associated_emails:
          type: array
          items:
            type: string
          description: "Associated email addresses — empty: requires data enrichment service"
        user_activity:
          type: string
          nullable: true
          description: "Activity level — null: requires live HLR/SS7 network query"
        leaked_online:
          type: boolean
          nullable: true
          description: "Whether number appears in breach databases — null: requires breach DB"
        reassigned:
          type: boolean
          nullable: true
          description: "Whether number was reassigned — null: requires FCC RND subscription"
      required:
        - valid
        - line_type
        - voip
        - carrier
        - country
        - city
        - region
        - timezones
        - active
        - fraud_score
        - fraud_reasons
        - recent_abuse
        - spammer
        - spam
        - prepaid
        - risky
        - dnc
        - dnc_source
        - pattern_flags
        - name
        - associated_emails
        - user_activity
        - leaked_online
        - reassigned
...