# lib/api-zod/src/generated/types/phoneLookupResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Phone intelligence result. Fields sourced from the phonenumbers library (valid, line_type, voip, carrier, country, city, region, timezones) are authoritative. Fields sourced from community datasets (spam, recent_abuse, spammer, dnc) reflect what has been reported. Fields requiring live carrier/breach/HLR access (name, associated_emails, user_activity, leaked_online, reassigned) are null — never fabricated.
 */
export interface PhoneLookupResult {
  /** Whether the number is a valid, dialable number per ITU standards */
  valid: boolean;
  /** Fixed Line, Mobile, VoIP, Toll-Free, Premium Rate, etc. */
  line_type: string;
  /** Whether the number is a VoIP line */
  voip: boolean;
  /** Original network carrier name (authoritative for most international numbers; may be empty for US mobile numbers due to number portability) */
  carrier: string;
  /** ISO 3166-1 alpha-2 country code */
  country: string;
  /** City or geographic area (from phonenumbers geocoder) */
  city: string;
  /** State, province, or region (from phonenumbers geocoder) */
  region: string;
  /** IANA timezone identifiers for this number's geographic area */
  timezones: string[];
  /** Best-effort heuristic — true if number is valid and not flagged in any community abuse dataset. NOT a live reachability check. */
  active: boolean;
  /** Risk score 0–100 computed from community data and pattern analysis */
  fraud_score: number;
  /** Reasons contributing to the fraud score */
  fraud_reasons: string[];
  /** True if number appears in community abuse/spam datasets */
  recent_abuse: boolean;
  /** True if number appears in community spam datasets */
  spammer: boolean;
  /** True if number appears in community spam datasets */
  spam: boolean;
  /** Heuristic — true if carrier name matches known prepaid brands */
  prepaid: boolean;
  /** True if fraud_score >= 75 or flagged in community datasets */
  risky: boolean;
  /** Community-proxy DNC flag — true if number appears in abuse reports (not the official FTC Do Not Call Registry, which requires subscription) */
  dnc: boolean;
  /** Source of DNC determination */
  dnc_source: string;
  /** Structural pattern risk indicators (e.g. 555-0xxx, sequential digits, high-risk area code) */
  pattern_flags: string[];
  /** Subscriber name — null: requires CNAM carrier lookup */
  name: string | null;
  /** Associated email addresses — empty: requires data enrichment service */
  associated_emails: string[];
  /** Activity level — null: requires live HLR/SS7 network query */
  user_activity: string | null;
  /** Whether number appears in breach databases — null: requires breach DB */
  leaked_online: boolean | null;
  /** Whether number was reassigned — null: requires FCC RND subscription */
  reassigned: boolean | null;
}